refactor: [패키지 변경] user 패키지 위치 변경- kww
