+++ spring-boot/src/main/java/com/tody/dayori/comment/controller/CommentController.java
@@ -7,9 +7,8 @@ import com.tody.dayori.comment.service.CommentService;
 import com.tody.dayori.common.dto.BaseResponse;
 import com.tody.dayori.page.domain.Page;
 import com.tody.dayori.page.service.PageService;
-import com.tody.dayori.user.domain.User;
+import com.tody.dayori.auth.entity.User;
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
