+++ spring-boot/src/main/java/com/tody/dayori/diary/domain/UserDiary.java
@@ -1,6 +1,6 @@
 package com.tody.dayori.diary.domain;
 
-import com.tody.dayori.user.domain.User;
+import com.tody.dayori.auth.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
