+++ spring-boot/src/main/java/com/tody/dayori/page/controller/PageController.java
@@ -8,8 +8,7 @@ import com.tody.dayori.page.dto.DeletePageRequest;
 import com.tody.dayori.page.dto.SearchPageRequest;
 import com.tody.dayori.page.dto.UpdatePageRequest;
 import com.tody.dayori.page.service.PageService;
-import com.tody.dayori.user.domain.User;
-import com.tody.dayori.user.service.UserService;
+import com.tody.dayori.auth.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
